Step 2.  Collapse renderer and parser into one directory. Update licences since it is now all under GPLv2 for simplicity. Bump version number to 0.17.

+++ test/renderercheck.sh
@@ -20,19 +20,23 @@
 # Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 #
 
-
 for F in `ls *.msc` ; do
     echo "$F"
-    $VALGRIND ../bin/mscgen -T png -o $F.s.png -i $F || exit $?
+    $VALGRIND ../bin/mscgen -T png -o $F.png -i $F || exit $?
     $VALGRIND ../bin/mscgen -T svg -o $F.svg -i $F || exit $?
     $VALGRIND ../bin/mscgen -T eps -o $F.eps -i $F || exit $?
 
-    # Try running script directly
-    ./$F
+    if [ -e /usr/bin/mscgen ] ; then
 
-    # Check direct running gave same output, then remove duplicate image
-    cmp ./$F.png  $F.s.png || exit $?
-    rm $F.s.png
+      # Try running script directly
+      mv $F.png $F.s.png
+      ./$F
+
+      # Check direct running gave same output, then remove duplicate image
+      cmp ./$F.png  $F.s.png || exit $?
+      rm $F.s.png
+    fi
+
 done
 
 # END OF SCRIPT
